Refactors Stories
